# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Baskonia Pipeline API",
    "description": "API de análisis de Baskonia (fase F3 de la migración strangler-fig).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/teams/{slug}/schedule-difficulty": {
      "get": {
        "tags": [
          "matchups"
        ],
        "summary": "Get Schedule Difficulty",
        "description": "Dificultad del próximo tramo de calendario (endpoint 12).",
        "operationId": "get_schedule_difficulty_api_v1_teams__slug__schedule_difficulty_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del equipo (p.ej. 'vitoria')",
              "title": "Slug"
            },
            "description": "Slug del equipo (p.ej. 'vitoria')"
          },
          {
            "name": "next_n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 5,
              "title": "Next N"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Año de inicio de temporada (p.ej. 2025)",
              "title": "Season"
            },
            "description": "Año de inicio de temporada (p.ej. 2025)"
          },
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Competición: acb | euroleague | supercopa",
              "title": "League"
            },
            "description": "Competición: acb | euroleague | supercopa"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDifficultyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teams/{slug}/narrative": {
      "get": {
        "tags": [
          "matchups"
        ],
        "summary": "Get Narrative",
        "description": "Narrativa de scouting en español (endpoint 13).",
        "operationId": "get_narrative_api_v1_teams__slug__narrative_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del equipo (p.ej. 'vitoria')",
              "title": "Slug"
            },
            "description": "Slug del equipo (p.ej. 'vitoria')"
          },
          {
            "name": "recent_n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 5,
              "title": "Recent N"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Año de inicio de temporada (p.ej. 2025)",
              "title": "Season"
            },
            "description": "Año de inicio de temporada (p.ej. 2025)"
          },
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Competición: acb | euroleague | supercopa",
              "title": "League"
            },
            "description": "Competición: acb | euroleague | supercopa"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NarrativeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teams/{slug}/matchups/{opponent_slug}/projection": {
      "get": {
        "tags": [
          "matchups"
        ],
        "summary": "Get Projection",
        "description": "Proyección de marcador entre dos equipos (endpoint 14).",
        "operationId": "get_projection_api_v1_teams__slug__matchups__opponent_slug__projection_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del equipo (p.ej. 'vitoria')",
              "title": "Slug"
            },
            "description": "Slug del equipo (p.ej. 'vitoria')"
          },
          {
            "name": "opponent_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del rival (p.ej. 'bilbao')",
              "title": "Opponent Slug"
            },
            "description": "Slug del rival (p.ej. 'bilbao')"
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Año de inicio de temporada (p.ej. 2025)",
              "title": "Season"
            },
            "description": "Año de inicio de temporada (p.ej. 2025)"
          },
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Competición: acb | euroleague | supercopa",
              "title": "League"
            },
            "description": "Competición: acb | euroleague | supercopa"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teams/{slug}/matchups/{opponent_slug}/head-to-head": {
      "get": {
        "tags": [
          "matchups"
        ],
        "summary": "Get Head To Head",
        "description": "Enfrentamientos directos entre dos equipos (endpoint 15).",
        "operationId": "get_head_to_head_api_v1_teams__slug__matchups__opponent_slug__head_to_head_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del equipo (p.ej. 'vitoria')",
              "title": "Slug"
            },
            "description": "Slug del equipo (p.ej. 'vitoria')"
          },
          {
            "name": "opponent_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del rival (p.ej. 'bilbao')",
              "title": "Opponent Slug"
            },
            "description": "Slug del rival (p.ej. 'bilbao')"
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Año de inicio de temporada (p.ej. 2025)",
              "title": "Season"
            },
            "description": "Año de inicio de temporada (p.ej. 2025)"
          },
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Competición: acb | euroleague | supercopa",
              "title": "League"
            },
            "description": "Competición: acb | euroleague | supercopa"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeadToHeadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DifficultyOpponent": {
        "properties": {
          "opponent_name": {
            "type": "string",
            "title": "Opponent Name"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "net_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Rating"
          }
        },
        "type": "object",
        "required": [
          "opponent_name",
          "date"
        ],
        "title": "DifficultyOpponent",
        "description": "Rival considerado en la dificultad de calendario."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HeadToHeadGame": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "league": {
            "type": "string",
            "title": "League"
          },
          "team_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Score"
          },
          "opponent_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Score"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "id",
          "date",
          "league"
        ],
        "title": "HeadToHeadGame",
        "description": "Un enfrentamiento directo entre dos equipos."
      },
      "HeadToHeadResponse": {
        "properties": {
          "team": {
            "$ref": "#/components/schemas/TeamRef"
          },
          "opponent": {
            "$ref": "#/components/schemas/TeamRef"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HeadToHeadGame"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "team",
          "opponent",
          "items"
        ],
        "title": "HeadToHeadResponse",
        "description": "Enfrentamientos directos entre dos equipos."
      },
...
      "NarrativeResponse": {
        "properties": {
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "league": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "League"
          },
          "recent_n": {
            "type": "integer",
            "title": "Recent N"
          },
          "narrative": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative"
          }
        },
        "type": "object",
        "required": [
          "season",
          "recent_n"
        ],
        "title": "NarrativeResponse",
        "description": "Narrativa de scouting (único campo en español de la API)."
      },
...
      "Projection": {
        "properties": {
          "projected_possessions": {
            "type": "number",
            "title": "Projected Possessions"
          },
          "team_projected_rating": {
            "type": "number",
            "title": "Team Projected Rating"
          },
          "opp_projected_rating": {
            "type": "number",
            "title": "Opp Projected Rating"
          },
          "team_projected_score": {
            "type": "number",
            "title": "Team Projected Score"
          },
          "opp_projected_score": {
            "type": "number",
            "title": "Opp Projected Score"
          },
          "expected_margin": {
            "type": "number",
            "title": "Expected Margin"
          }
        },
        "type": "object",
        "required": [
          "projected_possessions",
          "team_projected_rating",
          "opp_projected_rating",
          "team_projected_score",
          "opp_projected_score",
          "expected_margin"
        ],
        "title": "Projection",
        "description": "Proyección de marcador esperado entre dos equipos."
      },
      "ProjectionResponse": {
        "properties": {
          "team": {
            "$ref": "#/components/schemas/TeamRef"
          },
          "opponent": {
            "$ref": "#/components/schemas/TeamRef"
          },
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "projection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Projection"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "team",
          "opponent",
          "season"
        ],
        "title": "ProjectionResponse",
        "description": "Proyección de un enfrentamiento (projection puede ser null)."
      },
...
      "ScheduleDifficultyResponse": {
        "properties": {
          "games_considered": {
            "type": "integer",
            "title": "Games Considered"
          },
          "opponents_scouted": {
            "type": "integer",
            "title": "Opponents Scouted"
          },
          "avg_opponent_net_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Opponent Net Rating"
          },
          "league": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "League"
          },
          "opponents": {
            "items": {
              "$ref": "#/components/schemas/DifficultyOpponent"
            },
            "type": "array",
            "title": "Opponents"
          }
        },
        "type": "object",
        "required": [
          "games_considered",
          "opponents_scouted",
          "opponents"
        ],
        "title": "ScheduleDifficultyResponse",
        "description": "Dificultad del próximo tramo de calendario."
      },
...
      "TeamRef": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name"
        ],
        "title": "TeamRef",
        "description": "Referencia a un equipo (slug + nombre para mostrar)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}